Refactor test structure.
- refactor integration test structure.
