+++ tests/src/test/java/io/streamnative/pulsar/handlers/mqtt/mqtt3/hivemq/base/HiveMQIntegrationTest.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.streamnative.pulsar.handlers.mqtt.hivemq;
+package io.streamnative.pulsar.handlers.mqtt.mqtt3.hivemq.base;
 
 import com.hivemq.client.mqtt.MqttGlobalPublishFilter;
 import com.hivemq.client.mqtt.datatypes.MqttQos;
