+++ tests/src/test/java/io/streamnative/pulsar/handlers/mqtt/mqtt5/hivemq/proxy/MQTT5AuthorizationProxyReasonCodeOnAllAckTest.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.streamnative.pulsar.handlers.mqtt.mqtt5;
+package io.streamnative.pulsar.handlers.mqtt.mqtt5.hivemq.proxy;
 
 import com.hivemq.client.mqtt.MqttGlobalPublishFilter;
 import com.hivemq.client.mqtt.datatypes.MqttQos;
@@ -25,9 +25,12 @@ import com.hivemq.client.mqtt.mqtt5.message.publish.Mqtt5Publish;
 import com.hivemq.client.mqtt.mqtt5.message.publish.puback.Mqtt5PubAckReasonCode;
 import com.hivemq.client.mqtt.mqtt5.message.subscribe.suback.Mqtt5SubAck;
 import com.hivemq.client.mqtt.mqtt5.message.subscribe.suback.Mqtt5SubAckReasonCode;
+import io.streamnative.pulsar.handlers.mqtt.MQTTServerConfiguration;
 import io.streamnative.pulsar.handlers.mqtt.base.AuthorizationConfig;
+import io.streamnative.pulsar.handlers.mqtt.mqtt5.hivemq.base.MQTT5ClientUtils;
 import java.nio.charset.StandardCharsets;
 import java.util.HashSet;
+import java.util.Random;
 import java.util.Set;
 import org.apache.pulsar.client.admin.PulsarAdminException;
 import org.apache.pulsar.common.policies.data.AuthAction;
@@ -35,7 +38,15 @@ import org.awaitility.Awaitility;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-public class MQTT5AuthorizationReasonCodeOnAllAckTest extends AuthorizationConfig {
+public class MQTT5AuthorizationProxyReasonCodeOnAllAckTest extends AuthorizationConfig {
+    private final Random random = new Random();
+
+    @Override
+    public MQTTServerConfiguration initConfig() throws Exception {
+        MQTTServerConfiguration conf = super.initConfig();
+        conf.setMqttProxyEnabled(true);
+        return conf;
+    }
 
     @Test(timeOut = TIMEOUT)
     public void testSubScribeAuthorized() throws PulsarAdminException, InterruptedException {
@@ -51,8 +62,8 @@ public class MQTT5AuthorizationReasonCodeOnAllAckTest extends AuthorizationConfi
 
         String message = "Hello MQTT Pulsar";
 
-        Mqtt5BlockingClient publisher = MQTT5ClientUtils.createMqtt5Client(
-                mqttBrokerPortList.get(0));
+        Mqtt5BlockingClient publisher = MQTT5ClientUtils.createMqtt5ProxyClient(
+                getMqttProxyPortList().get(random.nextInt(mqttProxyPortList.size())));
         publisher.connectWith()
                 .simpleAuth()
                 .username("user1")
@@ -60,8 +71,8 @@ public class MQTT5AuthorizationReasonCodeOnAllAckTest extends AuthorizationConfi
                 .applySimpleAuth()
                 .send();
 
-        Mqtt5BlockingClient consumer = MQTT5ClientUtils.createMqtt5Client(
-                mqttBrokerPortList.get(0));
+        Mqtt5BlockingClient consumer = MQTT5ClientUtils.createMqtt5ProxyClient(
+                getMqttProxyPortList().get(random.nextInt(mqttProxyPortList.size())));
         consumer.connectWith()
                 .simpleAuth()
                 .username("user2")
@@ -85,7 +96,7 @@ public class MQTT5AuthorizationReasonCodeOnAllAckTest extends AuthorizationConfi
 
     @Test(timeOut = TIMEOUT)
     public void testSubScribeNotAuthorized() {
-        Mqtt5BlockingClient client = MQTT5ClientUtils.createMqtt5Client(getMqttBrokerPortList().get(0));
+        Mqtt5BlockingClient client = MQTT5ClientUtils.createMqtt5ProxyClient(getMqttBrokerPortList().get(0));
         client.connectWith()
                 .simpleAuth()
                 .username("user1")
@@ -105,7 +116,7 @@ public class MQTT5AuthorizationReasonCodeOnAllAckTest extends AuthorizationConfi
 
     @Test(timeOut = TIMEOUT)
     public void testPublishNotAuthorized() {
-        Mqtt5BlockingClient client = MQTT5ClientUtils.createMqtt5Client(getMqttBrokerPortList().get(0));
+        Mqtt5BlockingClient client = MQTT5ClientUtils.createMqtt5ProxyClient(getMqttBrokerPortList().get(0));
         client.connectWith()
                 .simpleAuth()
                 .username("user1")
@@ -130,7 +141,7 @@ public class MQTT5AuthorizationReasonCodeOnAllAckTest extends AuthorizationConfi
 
     @Test(timeOut = TIMEOUT)
     public void testAuthenticationFail() {
-        Mqtt5BlockingClient client = MQTT5ClientUtils.createMqtt5Client(getMqttBrokerPortList().get(0));
+        Mqtt5BlockingClient client = MQTT5ClientUtils.createMqtt5ProxyClient(getMqttBrokerPortList().get(0));
         try {
             client.connectWith()
                     .simpleAuth()
@@ -146,7 +157,7 @@ public class MQTT5AuthorizationReasonCodeOnAllAckTest extends AuthorizationConfi
 
     @Test(timeOut = TIMEOUT)
     public void testAuthenticationSuccess() {
-        Mqtt5BlockingClient client = MQTT5ClientUtils.createMqtt5Client(getMqttBrokerPortList().get(0));
+        Mqtt5BlockingClient client = MQTT5ClientUtils.createMqtt5ProxyClient(getMqttBrokerPortList().get(0));
         Mqtt5ConnAck connAck = client.connectWith()
                 .simpleAuth()
                 .username("user1")
@@ -156,3 +167,4 @@ public class MQTT5AuthorizationReasonCodeOnAllAckTest extends AuthorizationConfi
         Assert.assertEquals(connAck.getReasonCode(), Mqtt5ConnAckReasonCode.SUCCESS);
     }
 }
+
